data_provider/requirements.txt
python-dotenv
kafka-python
pandas
logging
